+++ packages/adhere/src/mobile-popover-menu/index.js
@@ -1,0 +1,2 @@
+import Model from '@baifendian/adhere-mobile-ui-popovermenu/es';
+export default Model;